add workaround for ath6kl driver
The ath6kl driver requires the interface to be brought up in order
for scanning to function. Without this change, the supplicant fails
on first activation of wifi, and will only connect successfully
after multiple retries. This workaround is also necessary to fix
Wi-Fi hotspot mode (in conjunction with another change to
system/netd).

Change-Id: Ia8c30bb780ae768cd723b279f69e596f5e09580b